daemon_config: Move proxy discover into the daemon_config
Based on review feedback, it makes more sense to put the proxy discovery into the DaemonConfig
code.
