+++ src/daemon/daemon_config.h
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2017 Canonical, Ltd.
+ * Copyright (C) 2017-2020 Canonical, Ltd.
  *
  * This program is free software; you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -34,6 +34,8 @@
 #include <multipass/vm_image_host.h>
 #include <multipass/vm_image_vault.h>
 
+#include <QNetworkProxy>
+
 #include <memory>
 #include <vector>
 
@@ -52,6 +54,7 @@ struct DaemonConfig
     const std::unique_ptr<CertStore> client_cert_store;
     const std::unique_ptr<UpdatePrompt> update_prompt;
     const std::shared_ptr<logging::MultiplexingLogger> logger;
+    const std::unique_ptr<QNetworkProxy> network_proxy;
     const multipass::Path cache_directory;
     const multipass::Path data_directory;
     const std::string server_address;
@@ -72,6 +75,7 @@ struct DaemonConfigBuilder
     std::unique_ptr<CertStore> client_cert_store;
     std::unique_ptr<UpdatePrompt> update_prompt;
     std::unique_ptr<logging::Logger> logger;
+    std::unique_ptr<QNetworkProxy> network_proxy;
     multipass::Path cache_directory;
     multipass::Path data_directory;
     std::string server_address;
