+++ src/network/url_downloader.cpp
@@ -109,39 +109,6 @@ QByteArray download(QNetworkAccessManager* manager, const Time& timeout, QUrl co
     }
     return reply->readAll();
 }
-
-std::unique_ptr<QNetworkProxy> discover_http_proxy()
-{
-    std::unique_ptr<QNetworkProxy> proxy_ptr{nullptr};
-
-    QString http_proxy{qgetenv("http_proxy")};
-    if (http_proxy.isEmpty())
-    {
-        // Some OS's are case senstive
-        http_proxy = qgetenv("HTTP_PROXY");
-    }
-
-    if (!http_proxy.isEmpty())
-    {
-        if (!http_proxy.startsWith("http://"))
-        {
-            http_proxy.prepend("http://");
-        }
-
-        QUrl proxy_url{http_proxy};
-        const auto host = proxy_url.host();
-        const auto port = proxy_url.port();
-
-        auto network_proxy = QNetworkProxy(QNetworkProxy::HttpProxy, host, static_cast<quint16>(port),
-                                           proxy_url.userName(), proxy_url.password());
-
-        QNetworkProxy::setApplicationProxy(network_proxy);
-
-        proxy_ptr = std::make_unique<QNetworkProxy>(network_proxy);
-    }
-
-    return proxy_ptr;
-}
 } // namespace
 
 mp::URLDownloader::URLDownloader(std::chrono::milliseconds timeout) : URLDownloader{Path(), timeout}
@@ -149,7 +116,7 @@ mp::URLDownloader::URLDownloader(std::chrono::milliseconds timeout) : URLDownloa
 }
 
 mp::URLDownloader::URLDownloader(const mp::Path& cache_dir, std::chrono::milliseconds timeout)
-    : cache_dir_path{QDir(cache_dir).filePath("network-cache")}, timeout{timeout}, network_proxy{discover_http_proxy()}
+    : cache_dir_path{QDir(cache_dir).filePath("network-cache")}, timeout{timeout}
 {
 }
